+++ edx_django_utils/monitoring/code_owner/utils.py
@@ -18,39 +18,45 @@ def get_code_owner_from_module(module):
     this lookup would match on 'openedx.features.discounts' before
     'openedx.features', because the former is more specific.
 
-    Uses CODE_OWNER_MAPPINGS Django Setting as detailed in:
+    See how to:
     https://github.com/edx/edx-django-utils/blob/master/edx_django_utils/monitoring/docs/how_tos/add_code_owner_custom_attribute_to_an_ida.rst
 
     """
-    assert _PATH_TO_CODE_OWNER_MAPPINGS != _INVALID_CODE_OWNER_MAPPING,\
-        'CODE_OWNER_MAPPINGS django setting set with invalid configuration. See logs for details.'
+    code_owner_mappings = get_code_owner_mappings()
+    if code_owner_mappings is None:
+        return None
 
     module_parts = module.split('.')
     # To make the most specific match, start with the max number of parts
     for number_of_parts in range(len(module_parts), 0, -1):
         partial_path = '.'.join(module_parts[0:number_of_parts])
-        if partial_path in _PATH_TO_CODE_OWNER_MAPPINGS:
-            code_owner = _PATH_TO_CODE_OWNER_MAPPINGS[partial_path]
+        if partial_path in code_owner_mappings:
+            code_owner = code_owner_mappings[partial_path]
             return code_owner
     return None
 
 
 def is_code_owner_mappings_configured():
     """
-    Returs True if code owner mappings were configured, and False otherwise.
+    Returns True if code owner mappings were configured, and False otherwise.
     """
-    return bool(_PATH_TO_CODE_OWNER_MAPPINGS)
+    return isinstance(get_code_owner_mappings(), dict)
 
 
-def _process_code_owner_mappings():
+# cached lookup table for code owner given a module path.
+# do not access this directly, but instead use get_code_owner_mappings.
+_PATH_TO_CODE_OWNER_MAPPINGS = None
+
+
+def get_code_owner_mappings():
     """
-    Processes the CODE_OWNER_MAPPINGS Django Setting and returns a dict optimized
+    Returns the contents of the CODE_OWNER_MAPPINGS Django Setting, processed
     for efficient lookup by path.
 
     Returns:
-         (dict): optimized dict for success processing, None if there are no
-            configured mappings, or _INVALID_CODE_OWNER_MAPPING if there is an
-            error processing the setting.
+         (dict): dict mapping modules to code owners, or None if there are no
+            configured mappings, or an empty dict if there is an error processing
+            the setting.
 
     Example return value::
 
@@ -61,48 +67,50 @@ def _process_code_owner_mappings():
         }
 
     """
-    _CODE_OWNER_MAPPINGS = getattr(settings, 'CODE_OWNER_MAPPINGS', None)
-    if not _CODE_OWNER_MAPPINGS:
+    global _PATH_TO_CODE_OWNER_MAPPINGS
+
+    # Return cached processed mappings if already processed
+    if _PATH_TO_CODE_OWNER_MAPPINGS is not None:
+        return _PATH_TO_CODE_OWNER_MAPPINGS
+
+    # Uses temporary variable to build mappings to avoid multi-threading issue with a partially
+    # processed map.  Worst case, it is processed more than once at start-up.
+    path_to_code_owner_mapping = {}
+
+    # .. setting_name: CODE_OWNER_MAPPINGS
+    # .. setting_default: None
+    # .. setting_description: Used for monitoring and reporting of ownership. Use a
+    #      dict with keys of code owner name and value as a list of dotted path
+    #      module names owned by the code owner.
+    code_owner_mappings = getattr(settings, 'CODE_OWNER_MAPPINGS', None)
+    if code_owner_mappings is None:
         return None
 
     try:
-        path_to_code_owner_mappings = {}
-        for code_owner in _CODE_OWNER_MAPPINGS:
-            path_list = _CODE_OWNER_MAPPINGS[code_owner]
+        for code_owner in code_owner_mappings:
+            path_list = code_owner_mappings[code_owner]
             for path in path_list:
-                path_to_code_owner_mappings[path] = code_owner
+                path_to_code_owner_mapping[path] = code_owner
                 optional_module_prefix_match = _OPTIONAL_MODULE_PREFIX_PATTERN.match(path)
                 # if path has an optional prefix, also add the module name without the prefix
                 if optional_module_prefix_match:
                     path_without_prefix = path[optional_module_prefix_match.end():]
-                    path_to_code_owner_mappings[path_without_prefix] = code_owner
-
-        return path_to_code_owner_mappings
+                    path_to_code_owner_mapping[path_without_prefix] = code_owner
     except Exception as e:  # pylint: disable=broad-except
         log.exception('Error processing code_owner_mappings. {}'.format(e))
-        # errors should be unlikely due do scripting the setting values.
-        # this will also trigger an error custom attribute that can be alerted on.
-        return _INVALID_CODE_OWNER_MAPPING
+
+    _PATH_TO_CODE_OWNER_MAPPINGS = path_to_code_owner_mapping
+    return _PATH_TO_CODE_OWNER_MAPPINGS
 
 
-# .. toggle_name: CODE_OWNER_MAPPINGS
-# .. toggle_implementation: DjangoSetting
-# .. toggle_default: None
-# .. toggle_description: Used to set monitoring custom attributes for owner. Dict with keys of code owner and value as
-#      list of dotted path module names owned by code owner.
-# .. toggle_category: monitoring
-# .. toggle_use_cases: open_edx
-# .. toggle_creation_date: 2020-05-29
-# .. toggle_expiration_date: None
-# .. toggle_tickets: None
-# .. toggle_status: supported
-# .. toggle_warnings: None
-# TODO: Update annotations.  This is more of a non-toggle setting than a toggle setting.
-_CODE_OWNER_MAPPINGS = None
+def clear_cached_mappings():
+    """
+    Clears the cached path to code owner mappings. Useful for testing.
+    """
+    global _PATH_TO_CODE_OWNER_MAPPINGS
+    _PATH_TO_CODE_OWNER_MAPPINGS = None
 
-# TODO: Remove this LMS spcific configuration by replacing with a Django Setting named
+
+# TODO: Remove this LMS specific configuration by replacing with a Django Setting named
 #    CODE_OWNER_OPTIONAL_MODULE_PREFIXES that takes a list of module prefixes (without the final period).
 _OPTIONAL_MODULE_PREFIX_PATTERN = re.compile(r'^(lms|common|openedx\.core)\.djangoapps\.')
-_INVALID_CODE_OWNER_MAPPING = 'invalid-code-owner-mapping'
-# lookup table for code owner given a module path
-_PATH_TO_CODE_OWNER_MAPPINGS = _process_code_owner_mappings()